BUG: Make nbagg work without IPython installed
