+++ lib/matplotlib/backends/backend_nbagg.py
@@ -26,9 +26,10 @@ from matplotlib.figure import Figure
 from matplotlib import is_interactive
 from matplotlib.backends.backend_webagg_core import (FigureManagerWebAgg,
                                                      FigureCanvasWebAggCore,
-                                                     NavigationToolbar2WebAgg)
+                                                     NavigationToolbar2WebAgg,
+                                                     TimerTornado)
 from matplotlib.backend_bases import (ShowBase, NavigationToolbar2,
-                                      TimerBase, FigureCanvasBase)
+                                      FigureCanvasBase)
 
 
 class Show(ShowBase):
@@ -183,38 +184,6 @@ class FigureManagerNbAgg(FigureManagerWebAgg):
     def remove_comm(self, comm_id):
         self.web_sockets = set([socket for socket in self.web_sockets
                                 if not socket.comm.comm_id == comm_id])
-
-
-class TimerTornado(TimerBase):
-    def _timer_start(self):
-        self._timer_stop()
-        if self._single:
-            ioloop = tornado.ioloop.IOLoop.instance()
-            self._timer = ioloop.add_timeout(
-                datetime.timedelta(milliseconds=self.interval),
-                self._on_timer)
-        else:
-            self._timer = tornado.ioloop.PeriodicCallback(
-                self._on_timer,
-                self.interval)
-            self._timer.start()
-
-    def _timer_stop(self):
-        if self._timer is None:
-            return
-        elif self._single:
-            ioloop = tornado.ioloop.IOLoop.instance()
-            ioloop.remove_timeout(self._timer)
-        else:
-            self._timer.stop()
-
-        self._timer = None
-
-    def _timer_set_interval(self):
-        # Only stop and restart it if the timer has already been started
-        if self._timer is not None:
-            self._timer_stop()
-            self._timer_start()
 
 
 class FigureCanvasNbAgg(FigureCanvasWebAggCore):
